Add empty check for order refunds

+++ includes/class-wc-payments-webhook-processing-service.php
@@ -205,11 +205,13 @@ class WC_Payments_Webhook_Processing_Service {
 
 		// Delete refund based on WCPay refund id.
 		$wc_refunds = $order->get_refunds();
-		foreach ( $wc_refunds as $wc_refund ) {
-			$wcpay_refund_id = $wc_refund->get_meta( '_wcpay_refund_id', true );
-			if ( $refund_id === $wcpay_refund_id ) {
-				$wc_refund->delete();
-				break;
+		if ( ! empty( $wc_refunds ) ) {
+			foreach ( $wc_refunds as $wc_refund ) {
+				$wcpay_refund_id = $wc_refund->get_meta( '_wcpay_refund_id', true );
+				if ( $refund_id === $wcpay_refund_id ) {
+					$wc_refund->delete();
+					break;
+				}
 			}
 		}
